Update the discord message format
Currently the discord report was sending 2 embeds. This
was an issue because if there was no test_results then we
would send the same file message twice unnecessarily.
This fix attempts to mediate the difference between
slack+teams messages with the discord in effort to keep the
reports standardised.

update the tests